Extract identical to own file
